# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/cars": {
            "get": {
                "description": "Find all car.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Find all car.",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Brand ID",
                        "name": "brand_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Model",
                        "name": "model",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum Year",
                        "name": "min_year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum Year",
                        "name": "max_year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebSuccess-array_response_CarResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebInternalServerError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Create a car.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Create car.",
                "parameters": [
                    {
                        "description": "the body to create a car",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CarCreateRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/web.WebSuccess-response_CarResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebBadRequestError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/web.WebForbiddenError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebInternalServerError"
                        }
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "get": {
                "description": "Find a car by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Find car.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Car ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebSuccess-response_CarResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.WebNotFoundError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebInternalServerError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a car.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Delete car.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Car ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebSuccess-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebBadRequestError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/web.WebForbiddenError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.WebNotFoundError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebInternalServerError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update a car.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Update car.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Car ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update a car",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CarUpdateRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebSuccess-response_CarResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebBadRequestError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/web.WebForbiddenError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.WebNotFoundError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebInternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CarCreateRequest": {
            "type": "object",
            "required": [
                "acceleration",
                "brand_id",
                "breaking_system_back",
                "breaking_system_front",
                "engine",
                "fuel",
                "height",
                "horse_power",
                "image_url",
                "length",
                "model",
                "torque",
                "transmission",
                "width",
                "year"
            ],
            "properties": {
                "brand_id": {
                    "type": "integer",
                    "x-order": "0"
                },
                "model": {
                    "type": "string",
                    "x-order": "1"
                },
                "acceleration": {
                    "type": "number",
                    "x-order": "10"
                },
                "horse_power": {
                    "type": "integer",
                    "x-order": "11"
                },
                "breaking_system_front": {
                    "type": "string",
                    "x-order": "12"
                },
                "breaking_system_back": {
                    "type": "string",
                    "x-order": "13"
                },
                "fuel": {
                    "type": "string",
                    "x-order": "14"
                },
                "year": {
                    "type": "integer",
                    "minimum": 1878,
                    "x-order": "2"
                },
                "image_url": {
                    "type": "string",
                    "x-order": "3"
                },
                "width": {
                    "type": "integer",
                    "x-order": "4"
                },
                "height": {
                    "type": "integer",
                    "x-order": "5"
                },
                "length": {
                    "type": "integer",
                    "x-order": "6"
                },
                "engine": {
                    "type": "string",
                    "x-order": "7"
                },
                "torque": {
                    "type": "integer",
                    "x-order": "8"
                },
                "transmission": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
        "request.CarUpdateRequest": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "integer",
                    "x-order": "0"
                },
                "model": {
                    "type": "string",
                    "x-order": "1"
                },
                "acceleration": {
                    "type": "number",
                    "x-order": "10"
                },
                "horse_power": {
                    "type": "integer",
                    "x-order": "11"
                },
                "breaking_system_front": {
                    "type": "string",
                    "x-order": "12"
                },
                "breaking_system_back": {
                    "type": "string",
                    "x-order": "13"
                },
                "fuel": {
                    "type": "string",
                    "x-order": "14"
                },
                "year": {
                    "type": "integer",
                    "minimum": 1878,
                    "x-order": "2"
                },
                "image_url": {
                    "type": "string",
                    "x-order": "3"
                },
                "width": {
                    "type": "integer",
                    "x-order": "4"
                },
                "height": {
                    "type": "integer",
                    "x-order": "5"
                },
                "length": {
                    "type": "integer",
                    "x-order": "6"
                },
                "engine": {
                    "type": "string",
                    "x-order": "7"
                },
                "torque": {
                    "type": "integer",
                    "x-order": "8"
                },
                "transmission": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
...
        "response.CarResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-order": "0",
                    "example": 1
                },
                "brand_id": {
                    "type": "integer",
                    "x-order": "1",
                    "example": 2
                },
                "transmission": {
                    "type": "string",
                    "x-order": "10",
                    "example": "Manual"
                },
                "acceleration": {
                    "type": "number",
                    "x-order": "11",
                    "example": 5.6
                },
                "horse_power": {
                    "type": "integer",
                    "x-order": "12",
                    "example": 265
                },
                "breaking_system_front": {
                    "type": "string",
                    "x-order": "13",
                    "example": "Ventilated Disc"
                },
                "breaking_system_back": {
                    "type": "string",
                    "x-order": "14",
                    "example": "Disc"
                },
                "fuel": {
                    "type": "string",
                    "x-order": "15",
                    "example": "Electric"
                },
                "model": {
                    "type": "string",
                    "x-order": "2",
                    "example": "Yaris"
                },
                "year": {
                    "type": "integer",
                    "x-order": "3",
                    "example": 2020
                },
                "image_url": {
                    "type": "string",
                    "x-order": "4",
                    "example": "image url"
                },
                "width": {
                    "type": "integer",
                    "x-order": "5",
                    "example": 462
                },
                "height": {
                    "type": "integer",
                    "x-order": "6",
                    "example": 184
                },
                "length": {
                    "type": "integer",
                    "x-order": "7",
                    "example": 137
                },
                "engine": {
                    "type": "string",
                    "x-order": "8",
                    "example": "2.0L EA113 CDLA TFSI In-Line 4 + Mild Hybrid 48V"
                },
                "torque": {
                    "type": "integer",
                    "x-order": "9",
                    "example": 370
                }
            }
        },
...
        "web.Metadata": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "x-order": "0"
                },
                "limit": {
                    "type": "integer",
                    "x-order": "1"
                },
                "total_pages": {
                    "type": "integer",
                    "x-order": "2"
                },
                "total_data": {
                    "type": "integer",
                    "x-order": "3"
                }
            }
        },
        "web.WebBadRequestError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "errors": {
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "web.WebForbiddenError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 403
                },
                "errors": {
                    "type": "string",
                    "example": "Forbidden"
                }
            }
        },
        "web.WebInternalServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "errors": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
        "web.WebNotFoundError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "errors": {
                    "type": "string",
                    "example": "Not Found"
                }
            }
        },
...
        "web.WebSuccess-array_response_CarResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "x-order": "0",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "x-order": "1",
                    "example": "success"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.CarResponse"
                    },
                    "x-order": "2"
                },
                "metadata": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/web.Metadata"
                        }
                    ],
                    "x-order": "3"
                }
            }
        },
...
        "web.WebSuccess-response_CarResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "x-order": "0",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "x-order": "1",
                    "example": "success"
                },
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.CarResponse"
                        }
                    ],
                    "x-order": "2"
                },
                "metadata": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/web.Metadata"
                        }
                    ],
                    "x-order": "3"
                }
            }
        },
...
        "web.WebSuccess-string": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "x-order": "0",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "x-order": "1",
                    "example": "success"
                },
                "data": {
                    "type": "string",
                    "x-order": "2"
                },
                "metadata": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/web.Metadata"
                        }
                    ],
                    "x-order": "3"
                }
            }
        },
...
    }
}